Extracted vote sorting into its own service since it will be used for both sending and receiving votes
