+++ packages/frontend/app/services/channels/vote-verifier.ts
@@ -1,11 +1,12 @@
 import Service, { inject as service } from '@ember/service';
 import VoteChain from 'emberclear/models/vote-chain';
-import Identity from 'emberclear/models/identity';
 import CryptoConnector from '../workers/crypto';
 import WorkersService from '../workers';
+import VoteSorter from 'emberclear/tests/unit/service/channels/vote-sorter';
 
 export default class VoteVerifier extends Service {
   @service workers!: WorkersService;
+  @service voteSorter!: VoteSorter;
   crypto?: CryptoConnector;
 
   async verify(voteToVerify: VoteChain): Promise<boolean> {
@@ -15,7 +16,7 @@ export default class VoteVerifier extends Service {
 
     this.connectCrypto();
 
-    let voteToVerifyActual: Uint8Array = this.generateSortedVote(voteToVerify);
+    let voteToVerifyActual: Uint8Array = this.voteSorter.generateSortedVote(voteToVerify);
     let voteToVerifyExpectedHash: Uint8Array = await this.crypto!.openSigned(
       voteToVerify.signature,
       voteToVerify.key.publicSigningKey
@@ -34,23 +35,6 @@ export default class VoteVerifier extends Service {
     this.crypto = new CryptoConnector({
       workerService: this.workers,
     });
-  }
-
-  private generateSortedVote(vote: VoteChain): Uint8Array {
-    let toReturn = [
-      this.toSortedPublicKeys(vote.remaining),
-      this.toSortedPublicKeys(vote.yes),
-      this.toSortedPublicKeys(vote.no),
-      vote.target.publicKey,
-      vote.action,
-      vote.key.publicKey,
-      vote.previousVoteChain.signature,
-    ];
-    return new TextEncoder().encode(JSON.stringify(toReturn));
-  }
-
-  private toSortedPublicKeys(identities: Identity[]): Uint8Array[] {
-    return identities.map((identity) => identity.publicKey).sort();
   }
 }
